+++ pkg/client/client.go
@@ -17,6 +17,7 @@ import (
 	v0040 "github.com/SlinkyProject/slurm-client/pkg/client/api/v0040"
 	v0041 "github.com/SlinkyProject/slurm-client/pkg/client/api/v0041"
 	v0042 "github.com/SlinkyProject/slurm-client/pkg/client/api/v0042"
+	v0043 "github.com/SlinkyProject/slurm-client/pkg/client/api/v0043"
 	"github.com/SlinkyProject/slurm-client/pkg/event"
 	"github.com/SlinkyProject/slurm-client/pkg/object"
 	"github.com/SlinkyProject/slurm-client/pkg/types"
@@ -61,6 +62,7 @@ type client struct {
 	v0040Client v0040.ClientInterface
 	v0041Client v0041.ClientInterface
 	v0042Client v0042.ClientInterface
+	v0043Client v0043.ClientInterface
 
 	server          string
 	authToken       string
@@ -94,11 +96,17 @@ func NewClient(config *Config, opts ...ClientOption) (Client, error) {
 		return nil, fmt.Errorf("unable to create client: %v", err)
 	}
 
+	v0043Client, err := v0043.NewSlurmClient(config.Server, config.AuthToken, config.HTTPClient)
+	if err != nil {
+		return nil, fmt.Errorf("unable to create client: %v", err)
+	}
+
 	// create return client object
 	client := &client{
 		v0040Client:     v0040Client,
 		v0041Client:     v0041Client,
 		v0042Client:     v0042Client,
+		v0043Client:     v0043Client,
 		informers:       make(map[object.ObjectType]InformerCache),
 		uncached:        make(set.Set[object.ObjectType]),
 		authToken:       config.AuthToken,
@@ -164,6 +172,20 @@ func (c *client) Create(
 		key := object.ObjectKey(fmt.Sprintf("%d", *jobId))
 		return c.Get(ctx, key, o)
 
+	case *types.V0043JobInfo:
+		var jobId *int32
+		var err error
+		if options.Allocation {
+			jobId, err = c.v0043Client.CreateJobInfoAlloc(ctx, req)
+		} else {
+			jobId, err = c.v0043Client.CreateJobInfo(ctx, req)
+		}
+		if err != nil {
+			return err
+		}
+		key := object.ObjectKey(fmt.Sprintf("%d", *jobId))
+		return c.Get(ctx, key, o)
+
 	default:
 		return errors.New(http.StatusText(http.StatusNotImplemented))
 	}
@@ -195,6 +217,11 @@ func (c *client) Delete(
 		return c.v0042Client.DeleteJobInfo(ctx, key)
 	case *types.V0042Node:
 		return c.v0042Client.DeleteNode(ctx, key)
+
+	case *types.V0043JobInfo:
+		return c.v0043Client.DeleteJobInfo(ctx, key)
+	case *types.V0043Node:
+		return c.v0043Client.DeleteNode(ctx, key)
 
 	default:
 		return errors.New(http.StatusText(http.StatusNotImplemented))
@@ -248,6 +275,19 @@ func (c *client) Update(
 		return c.Get(ctx, obj.GetKey(), o)
 	case *types.V0042Node:
 		err := c.v0042Client.UpdateNode(ctx, key, req)
+		if err != nil {
+			return err
+		}
+		return c.Get(ctx, obj.GetKey(), o)
+
+	case *types.V0043JobInfo:
+		err := c.v0043Client.UpdateJobInfo(ctx, key, req)
+		if err != nil {
+			return err
+		}
+		return c.Get(ctx, obj.GetKey(), o)
+	case *types.V0043Node:
+		err := c.v0043Client.UpdateNode(ctx, key, req)
 		if err != nil {
 			return err
 		}
@@ -366,6 +406,31 @@ func (c *client) Get(
 		}
 		*o = *out
 
+	case *types.V0043ControllerPing:
+		out, err := c.v0043Client.GetControllerPing(ctx, string(key))
+		if err != nil {
+			return err
+		}
+		*o = *out
+	case *types.V0043JobInfo:
+		out, err := c.v0043Client.GetJobInfo(ctx, string(key))
+		if err != nil {
+			return err
+		}
+		*o = *out
+	case *types.V0043Node:
+		out, err := c.v0043Client.GetNode(ctx, string(key))
+		if err != nil {
+			return err
+		}
+		*o = *out
+	case *types.V0043PartitionInfo:
+		out, err := c.v0043Client.GetPartitionInfo(ctx, string(key))
+		if err != nil {
+			return err
+		}
+		*o = *out
+
 	default:
 		return errors.New(http.StatusText(http.StatusNotImplemented))
 	}
@@ -476,6 +541,31 @@ func (c *client) List(
 		*objList = *out
 	case *types.V0042StatsList:
 		out, err := c.v0042Client.ListStats(ctx)
+		if err != nil {
+			return err
+		}
+		*objList = *out
+
+	case *types.V0043ControllerPingList:
+		out, err := c.v0043Client.ListControllerPing(ctx)
+		if err != nil {
+			return err
+		}
+		*objList = *out
+	case *types.V0043JobInfoList:
+		out, err := c.v0043Client.ListJobInfo(ctx)
+		if err != nil {
+			return err
+		}
+		*objList = *out
+	case *types.V0043NodeList:
+		out, err := c.v0043Client.ListNodes(ctx)
+		if err != nil {
+			return err
+		}
+		*objList = *out
+	case *types.V0043PartitionInfoList:
+		out, err := c.v0043Client.ListPartitionInfo(ctx)
 		if err != nil {
 			return err
 		}