+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to this project will be documented in this file.
 - Added v42 client library support.
 - Added v41, v42 support for sdiag statistics.
 - Added v43 OpenAPI spec and generated API.
+- Added v43 client library support.
 
 ### Fixed